@@ -13,42 +13,42 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "5.0.0",
-    "@angular/cdk": "5.0.0-rc0",
-    "@angular/common": "5.0.0",
-    "@angular/compiler": "5.0.0",
-    "@angular/core": "5.0.0",
-    "@angular/forms": "5.0.0",
-    "@angular/material": "5.0.0-rc0",
-    "@angular/platform-browser": "5.0.0",
-    "@angular/platform-browser-dynamic": "5.0.0",
-    "@angular/router": "5.0.0",
-    "angular-oauth2-oidc": "3.0.2",
-    "core-js": "2.5.1",
+    "@angular/animations": "5.1.0",
+    "@angular/cdk": "5.0.1",
+    "@angular/common": "5.1.0",
+    "@angular/compiler": "5.1.0",
+    "@angular/core": "5.1.0",
+    "@angular/forms": "5.1.0",
+    "@angular/material": "5.0.1",
+    "@angular/platform-browser": "5.1.0",
+    "@angular/platform-browser-dynamic": "5.1.0",
+    "@angular/router": "5.1.0",
+    "angular-oauth2-oidc": "3.0.3",
+    "core-js": "2.5.3",
     "hammerjs": "2.0.8",
-    "rxjs": "5.5.2",
+    "rxjs": "5.5.5",
     "web-animations-js": "2.3.1",
     "zone.js": "0.8.18"
   },
   "devDependencies": {
-    "@angular/cli": "1.5.0",
-    "@angular/compiler-cli": "5.0.0",
-    "@angular/language-service": "5.0.0",
-    "@types/jasmine": "2.6.2",
+    "@angular/cli": "1.6.0",
+    "@angular/compiler-cli": "5.1.0",
+    "@angular/language-service": "5.1.0",
+    "@types/jasmine": "2.8.2",
     "@types/jasminewd2": "2.0.3",
-    "@types/node": "8.0.47",
-    "codelyzer": "4.0.1",
+    "@types/node": "8.0.58",
+    "codelyzer": "4.0.2",
     "jasmine-core": "2.8.0",
     "jasmine-spec-reporter": "4.2.1",
     "karma": "1.7.1",
     "karma-chrome-launcher": "2.2.0",
     "karma-cli": "1.0.1",
     "karma-coverage-istanbul-reporter": "1.3.0",
-    "karma-jasmine": "1.1.0",
+    "karma-jasmine": "1.1.1",
     "karma-jasmine-html-reporter": "0.2.2",
-    "protractor": "5.2.0",
-    "ts-node": "3.3.0",
+    "protractor": "5.2.2",
+    "ts-node": "4.0.1",
     "tslint": "5.8.0",
-    "typescript": "2.6.1"
+    "typescript": "2.6.2"
   }
 }
